requirements.txt
streamlit
flask
google-generativeai
toml
python-dotenv